re-added older xmls, added "party-tag" in fpml section of main xml, added JAXB tests.
